requirements.txt
pytest==7.2.2
pytest-playwright==0.3.1
